requirements.txt
anyio==4.7.0
arrow==1.3.0
asgiref==3.8.1
attrs==23.2.0
Babel==2.10.3
bcc==0.29.1
bcrypt==3.2.2
binaryornot==0.4.4
blinker==1.7.0
Brlapi==0.8.5
Brotli==1.1.0
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
cloud-init==24.3.1
colorama==0.4.6
command-not-found==0.3
configobj==5.0.8
cookiecutter==2.6.0
cryptography==41.0.7
cupshelpers==1.0
dbus-python==1.3.2
defer==1.0.6
distro==1.9.0
distro-info==1.7+build1
Django==5.1.4
django-cors-headers==4.6.0
django-jazzmin==3.0.1
django-redis==5.4.0
djangorestframework==3.15.2
duplicity==2.1.4
fasteners==0.18
flet==0.25.1
flet-cli==0.25.1
flet-desktop==0.25.1
gyp==0.1
h11==0.14.0
httpcore==1.0.7
httplib2==0.20.4
httpx==0.28.1
idna==3.10
Jinja2==3.1.4
jsonpatch==1.32
jsonpointer==2.0
jsonschema==4.10.3
language-selector==0.1
launchpadlib==1.11.0
lazr.restfulclient==0.14.6
lazr.uri==1.0.6
louis==3.29.0
Mako==1.3.2.dev0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
monotonic==1.6
mutagen==1.46.0
netaddr==0.8.0
netifaces==0.11.0
oauthlib==3.2.2
olefile==0.46
packaging==24.2
paramiko==2.12.0
pexpect==4.9.0
pillow==10.2.0
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pycairo==1.25.1
pycryptodomex==3.20.0
pycups==2.0.1
Pygments==2.18.0
PyGObject==3.48.2
PyJWT==2.7.0
PyNaCl==1.5.0
pyparsing==3.1.1
pypng==0.20220715.0
pyrsistent==0.20.0
pyserial==3.5
python-apt==2.7.7+ubuntu3
python-dateutil==2.9.0.post0
python-debian==0.1.49+ubuntu2
python-slugify==8.0.4
pytz==2024.1
pyxdg==0.28
PyYAML==6.0.2
qrcode==7.4.2
redis==5.2.1
repath==0.9.0
requests==2.32.3
rich==13.9.4
screen-resolution-extra==0.0.0
setuptools==68.1.2
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.2
systemd-python==235
text-unidecode==1.3
toml==0.10.2
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.2
unattended-upgrades==0.1
urllib3==2.2.3
usb-creator==0.3.16
wadllib==1.3.6
watchdog==4.0.2
websockets==10.4
wheel==0.42.0
xdg==5
xkit==0.0.0
yt-dlp==2024.4.9
